HSEARCH-3439 Remove other unnecessary interfaces in implementation packages
