+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/schema/management/impl/ElasticsearchSchemaCreator.java
@@ -9,15 +9,27 @@ package org.hibernate.search.backend.elasticsearch.schema.management.impl;
 import java.util.concurrent.CompletableFuture;
 
 import org.hibernate.search.backend.elasticsearch.index.layout.impl.IndexNames;
-import org.hibernate.search.backend.elasticsearch.work.result.impl.ExistingIndexMetadata;
+import org.hibernate.search.backend.elasticsearch.index.layout.IndexLayoutStrategy;
 import org.hibernate.search.backend.elasticsearch.lowlevel.index.impl.IndexMetadata;
+import org.hibernate.search.backend.elasticsearch.util.spi.URLEncodedString;
+import org.hibernate.search.backend.elasticsearch.work.result.impl.ExistingIndexMetadata;
 import org.hibernate.search.util.common.SearchException;
 
 /**
  * An object responsible for creating an index and its mappings based on provided metadata.
- *
+ * @author Gunnar Morling
  */
-public interface ElasticsearchSchemaCreator {
+final class ElasticsearchSchemaCreator {
+
+	private final ElasticsearchSchemaAccessor schemaAccessor;
+
+	private final IndexLayoutStrategy indexLayoutStrategy;
+
+	public ElasticsearchSchemaCreator(ElasticsearchSchemaAccessor schemaAccessor,
+			IndexLayoutStrategy indexLayoutStrategy) {
+		this.schemaAccessor = schemaAccessor;
+		this.indexLayoutStrategy = indexLayoutStrategy;
+	}
 
 	/**
 	 * Create an index and its mapping.
@@ -27,7 +39,14 @@ public interface ElasticsearchSchemaCreator {
 	 * @return A future.
 	 * @throws SearchException If an error occurs.
 	 */
-	CompletableFuture<?> createIndexAssumeNonExisting(IndexNames indexNames, IndexMetadata indexMetadata);
+	public CompletableFuture<?> createIndexAssumeNonExisting(IndexNames indexNames, IndexMetadata indexMetadata) {
+		return schemaAccessor.createIndexAssumeNonExisting(
+				createPrimaryIndexName( indexNames ),
+				indexMetadata.getAliases(),
+				indexMetadata.getSettings(),
+				indexMetadata.getMapping()
+		);
+	}
 
 	/**
 	 * Create an index and its mapping, but only if the index doesn't already exist.
@@ -37,6 +56,28 @@ public interface ElasticsearchSchemaCreator {
 	 * @return A future holding the metadata of the pre-existing index, or null if the index had to be created.
 	 * @throws SearchException If an error occurs.
 	 */
-	CompletableFuture<ExistingIndexMetadata> createIndexIfAbsent(IndexNames indexNames, IndexMetadata indexMetadata);
+	public CompletableFuture<ExistingIndexMetadata> createIndexIfAbsent(IndexNames indexNames, IndexMetadata indexMetadata) {
+		return schemaAccessor.getCurrentIndexMetadataOrNull( indexNames )
+				.thenCompose( existingIndexMetadata -> {
+					if ( existingIndexMetadata != null ) {
+						return CompletableFuture.completedFuture( existingIndexMetadata );
+					}
+					else {
+						return schemaAccessor.createIndexIgnoreExisting(
+								createPrimaryIndexName( indexNames ),
+								indexMetadata.getAliases(),
+								indexMetadata.getSettings(),
+								indexMetadata.getMapping()
+						)
+								.thenApply( ignored -> null );
+					}
+				} );
+	}
+
+	private URLEncodedString createPrimaryIndexName(IndexNames indexNames) {
+		return IndexNames.encodeName(
+				indexLayoutStrategy.createInitialElasticsearchIndexName( indexNames.hibernateSearchIndex() )
+		);
+	}
 
 }
