+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/schema/management/impl/ElasticsearchSchemaValidator.java
@@ -6,14 +6,32 @@
  */
 package org.hibernate.search.backend.elasticsearch.schema.management.impl;
 
+import org.hibernate.search.backend.elasticsearch.lowlevel.index.aliases.impl.IndexAliasDefinition;
 import org.hibernate.search.backend.elasticsearch.lowlevel.index.impl.IndexMetadata;
+import org.hibernate.search.backend.elasticsearch.lowlevel.index.mapping.impl.RootTypeMapping;
+import org.hibernate.search.backend.elasticsearch.validation.impl.ElasticsearchValidationMessages;
+import org.hibernate.search.backend.elasticsearch.validation.impl.IndexAliasDefinitionValidator;
+import org.hibernate.search.backend.elasticsearch.validation.impl.IndexSettingsValidator;
+import org.hibernate.search.backend.elasticsearch.validation.impl.RootTypeMappingValidator;
+import org.hibernate.search.backend.elasticsearch.validation.impl.ValidationContextType;
+import org.hibernate.search.backend.elasticsearch.validation.impl.ValidationErrorCollector;
+import org.hibernate.search.backend.elasticsearch.validation.impl.Validator;
+import org.hibernate.search.backend.elasticsearch.lowlevel.index.settings.impl.IndexSettings;
 import org.hibernate.search.engine.reporting.spi.ContextualFailureCollector;
 
 /**
  * An object responsible for validating type mappings retrieved from an existing Elasticsearch instance
  * against Hibernate Search-generated type mappings.
+ * <p>
+ * <strong>Important implementation note:</strong> unexpected attributes (i.e. those not mapped to a field in TypeMapping)
+ * are totally ignored. This allows users to leverage Elasticsearch features that are not supported in
+ * Hibernate Search, by setting those attributes manually.
  */
-public interface ElasticsearchSchemaValidator {
+final class ElasticsearchSchemaValidator {
+
+	private final Validator<IndexAliasDefinition> aliasDefinitionValidator = new IndexAliasDefinitionValidator();
+	private final Validator<IndexSettings> indexSettingsValidator = new IndexSettingsValidator();
+	private final Validator<RootTypeMapping> rootTypeMappingValidator = new RootTypeMappingValidator();
 
 	/**
 	 * Validate actual index metadata, pushing any validation failure to the given collector.
@@ -22,9 +40,16 @@ public interface ElasticsearchSchemaValidator {
 	 * @param actualIndexMetadata The actual metadata, retrieved from Elasticsearch.
 	 * @param contextualFailureCollector A collector for validation failures.
 	 */
-	void validate(IndexMetadata expectedIndexMetadata,
-			IndexMetadata actualIndexMetadata,
-			ContextualFailureCollector contextualFailureCollector);
+	public void validate(IndexMetadata expectedIndexMetadata, IndexMetadata actualIndexMetadata,
+			ContextualFailureCollector contextualFailureCollector) {
+		ValidationErrorCollector errorCollector = new ValidationErrorCollector( contextualFailureCollector );
+
+		validateAliases( errorCollector, expectedIndexMetadata, actualIndexMetadata );
+
+		validateSettings( errorCollector, expectedIndexMetadata, actualIndexMetadata );
+
+		validateMapping( errorCollector, expectedIndexMetadata, actualIndexMetadata );
+	}
 
 	/**
 	 * Validate index aliases only, returning a boolean indicating
@@ -34,7 +59,20 @@ public interface ElasticsearchSchemaValidator {
 	 * @param actualIndexMetadata The actual metadata, retrieved from Elasticsearch.
 	 * @return {@code true} if the actual aliases are valid, {@code false} otherwise.
 	 */
-	boolean isAliasesValid(IndexMetadata expectedIndexMetadata, IndexMetadata actualIndexMetadata);
+	public boolean isAliasesValid(IndexMetadata expectedIndexMetadata, IndexMetadata actualIndexMetadata) {
+		ValidationErrorCollector errorCollector = new ValidationErrorCollector();
+		validateAliases( errorCollector, expectedIndexMetadata, actualIndexMetadata );
+		return !errorCollector.hasError();
+	}
+
+	private void validateAliases(ValidationErrorCollector errorCollector, IndexMetadata expectedIndexMetadata,
+			IndexMetadata actualIndexMetadata) {
+		aliasDefinitionValidator.validateAllIgnoreUnexpected(
+				errorCollector, ValidationContextType.ALIAS,
+				ElasticsearchValidationMessages.INSTANCE.aliasMissing(),
+				expectedIndexMetadata.getAliases(), actualIndexMetadata.getAliases()
+		);
+	}
 
 	/**
 	 * Validate index settings only, returning a boolean indicating
@@ -44,7 +82,20 @@ public interface ElasticsearchSchemaValidator {
 	 * @param actualIndexMetadata The actual metadata, retrieved from Elasticsearch.
 	 * @return {@code true} if the actual settings are valid, {@code false} otherwise.
 	 */
-	boolean isSettingsValid(IndexMetadata expectedIndexMetadata, IndexMetadata actualIndexMetadata);
+	public boolean isSettingsValid(IndexMetadata expectedIndexMetadata, IndexMetadata actualIndexMetadata) {
+		ValidationErrorCollector errorCollector = new ValidationErrorCollector();
+
+		validateSettings( errorCollector, expectedIndexMetadata, actualIndexMetadata );
+
+		return !errorCollector.hasError();
+	}
+
+	private void validateSettings(ValidationErrorCollector errorCollector,
+			IndexMetadata expectedIndexMetadata, IndexMetadata actualIndexMetadata) {
+		indexSettingsValidator.validate(
+				errorCollector, expectedIndexMetadata.getSettings(), actualIndexMetadata.getSettings()
+		);
+	}
 
 	/**
 	 * Validate index mapping only, returning a boolean indicating
@@ -54,6 +105,18 @@ public interface ElasticsearchSchemaValidator {
 	 * @param actualIndexMetadata The actual metadata, retrieved from Elasticsearch.
 	 * @return {@code true} if the actual mapping is valid, {@code false} otherwise.
 	 */
-	boolean isMappingValid(IndexMetadata expectedIndexMetadata, IndexMetadata actualIndexMetadata);
+	public boolean isMappingValid(IndexMetadata expectedIndexMetadata, IndexMetadata actualIndexMetadata) {
+		ValidationErrorCollector errorCollector = new ValidationErrorCollector();
+
+		validateMapping( errorCollector, expectedIndexMetadata, actualIndexMetadata );
+
+		return !errorCollector.hasError();
+	}
+
+	private void validateMapping(ValidationErrorCollector errorCollector,
+			IndexMetadata expectedIndexMetadata, IndexMetadata actualIndexMetadata) {
+		rootTypeMappingValidator.validate( errorCollector, expectedIndexMetadata.getMapping(),
+				actualIndexMetadata.getMapping() );
+	}
 
 }
