+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/work/impl/RefreshWork.java
@@ -31,7 +31,7 @@ public class RefreshWork extends AbstractNonBulkableWork<Void> {
 		private final List<URLEncodedString> indexNames = new ArrayList<>();
 
 		public Builder() {
-			super( DefaultElasticsearchRequestSuccessAssessor.INSTANCE );
+			super( ElasticsearchRequestSuccessAssessor.DEFAULT_INSTANCE );
 		}
 
 		public Builder index(URLEncodedString indexName) {
