+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/work/impl/LuceneWorkFactory.java
@@ -6,46 +6,104 @@
  */
 package org.hibernate.search.backend.lucene.work.impl;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Set;
 
 import org.hibernate.search.backend.lucene.document.impl.LuceneIndexEntry;
+import org.hibernate.search.backend.lucene.lowlevel.common.impl.MetadataFields;
+import org.hibernate.search.backend.lucene.lowlevel.query.impl.Queries;
+import org.hibernate.search.backend.lucene.multitenancy.impl.MultiTenancyStrategy;
 
 import org.apache.lucene.search.Explanation;
+import org.apache.lucene.search.MatchAllDocsQuery;
 import org.apache.lucene.search.Query;
 
+public final class LuceneWorkFactory {
 
-public interface LuceneWorkFactory {
+	private final MultiTenancyStrategy multiTenancyStrategy;
 
-	IndexManagementWork<Void> createIndexIfMissing();
+	public LuceneWorkFactory(MultiTenancyStrategy multiTenancyStrategy) {
+		this.multiTenancyStrategy = multiTenancyStrategy;
+	}
 
-	IndexManagementWork<Void> dropIndexIfExisting();
+	public IndexManagementWork<Void> createIndexIfMissing() {
+		return new CreateIndexIfMissingWork();
+	}
 
-	IndexManagementWork<Void> validateIndexExists();
+	public IndexManagementWork<Void> dropIndexIfExisting() {
+		return new DropIndexIfExistingWork();
+	}
 
-	IndexManagementWork<?> flush();
+	public IndexManagementWork<Void> validateIndexExists() {
+		return new ValidateIndexExistsWork();
+	}
 
-	IndexManagementWork<?> refresh();
+	public IndexManagementWork<?> flush() {
+		return new FlushWork();
+	}
 
-	IndexManagementWork<?> mergeSegments();
+	public IndexManagementWork<?> refresh() {
+		return new RefreshWork();
+	}
 
-	IndexManagementWork<Long> computeSizeInBytes();
+	public IndexManagementWork<?> mergeSegments() {
+		return new MergeSegmentsWork();
+	}
 
-	SingleDocumentIndexingWork add(String tenantId, String entityTypeName, Object entityIdentifier,
-			String documentIdentifier, LuceneIndexEntry indexEntry);
+	public IndexManagementWork<Long> computeSizeInBytes() {
+		return new ComputeSizeInBytesWork();
+	}
 
-	SingleDocumentIndexingWork update(String tenantId, String entityTypeName, Object entityIdentifier,
-			String documentIdentifier, LuceneIndexEntry indexEntry);
+	public SingleDocumentIndexingWork add(String tenantId, String entityTypeName, Object entityIdentifier,
+			String documentIdentifier, LuceneIndexEntry indexEntry) {
+		return new AddEntryWork( tenantId, entityTypeName, entityIdentifier,
+				documentIdentifier, indexEntry );
+	}
 
-	SingleDocumentIndexingWork delete(String tenantId, String entityTypeName, Object entityIdentifier, String id);
+	public SingleDocumentIndexingWork update(String tenantId, String entityTypeName, Object entityIdentifier,
+			String documentIdentifier, LuceneIndexEntry indexEntry) {
+		Query filter = multiTenancyStrategy.filterOrNull( tenantId );
+		return new UpdateEntryWork( tenantId, entityTypeName, entityIdentifier,
+				documentIdentifier, filter, indexEntry );
+	}
 
-	IndexManagementWork<?> deleteAll(String tenantId, Set<String> routingKeys);
+	public SingleDocumentIndexingWork delete(String tenantId, String entityTypeName, Object entityIdentifier,
+			String documentIdentifier) {
+		Query filter = multiTenancyStrategy.filterOrNull( tenantId );
+		return new DeleteEntryWork( tenantId, entityTypeName, entityIdentifier, documentIdentifier, filter );
+	}
 
-	<R> ReadWork<R> search(LuceneSearcher<R, ?> searcher, Integer offset, Integer limit, int totalHitCountThreshold);
+	public IndexManagementWork<?> deleteAll(String tenantId, Set<String> routingKeys) {
+		List<Query> filters = new ArrayList<>();
+		Query filter = multiTenancyStrategy.filterOrNull( tenantId );
+		if ( filter != null ) {
+			filters.add( filter );
+		}
+		if ( !routingKeys.isEmpty() ) {
+			filters.add( Queries.anyTerm( MetadataFields.routingKeyFieldName(), routingKeys ) );
+		}
 
-	<ER> ReadWork<ER> scroll(LuceneSearcher<?, ER> searcher, int offset, int limit, int totalHitCountThreshold);
+		return new DeleteEntriesByQueryWork( Queries.boolFilter( new MatchAllDocsQuery(), filters ) );
+	}
 
-	ReadWork<Integer> count(LuceneSearcher<?, ?> searcher);
+	public <R> ReadWork<R> search(LuceneSearcher<R, ?> searcher, Integer offset, Integer limit, int totalHitCountThreshold) {
+		return new SearchWork<>( searcher, offset, limit, totalHitCountThreshold );
+	}
 
-	ReadWork<Explanation> explain(LuceneSearcher<?, ?> searcher,
-			String explainedDocumentTypeName, String explainedDocumentId, Query explainedDocumentFilter);
+	public <ER> ReadWork<ER> scroll(LuceneSearcher<?, ER> searcher, int offset, int limit, int totalHitCountThreshold) {
+		return new ScrollWork<>( searcher, offset, limit, totalHitCountThreshold );
+	}
+
+	public ReadWork<Integer> count(LuceneSearcher<?, ?> searcher) {
+		return new CountWork( searcher );
+	}
+
+	public ReadWork<Explanation> explain(LuceneSearcher<?, ?> searcher,
+			String explainedDocumentTypeName, String explainedDocumentId, Query explainedDocumentFilter) {
+		return new ExplainWork(
+				searcher,
+				explainedDocumentTypeName, explainedDocumentId, explainedDocumentFilter
+		);
+	}
 }
