+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/mapping/building/impl/PojoIndexModelBinder.java
@@ -9,8 +9,11 @@ package org.hibernate.search.mapper.pojo.mapping.building.impl;
 import java.util.Map;
 import java.util.Optional;
 
+import org.hibernate.search.engine.environment.bean.BeanResolver;
 import org.hibernate.search.engine.mapper.mapping.building.spi.IndexBindingContext;
+import org.hibernate.search.engine.mapper.mapping.building.spi.IndexFieldTypeDefaultsProvider;
 import org.hibernate.search.engine.mapper.mapping.building.spi.IndexedEntityBindingContext;
+import org.hibernate.search.engine.mapper.mapping.building.spi.MappingBuildContext;
 import org.hibernate.search.mapper.pojo.bridge.binding.PropertyBindingContext;
 import org.hibernate.search.mapper.pojo.bridge.binding.TypeBindingContext;
 import org.hibernate.search.mapper.pojo.bridge.binding.ValueBindingContext;
@@ -18,18 +21,32 @@ import org.hibernate.search.mapper.pojo.bridge.binding.impl.BoundIdentifierBridg
 import org.hibernate.search.mapper.pojo.bridge.binding.impl.BoundPropertyBridge;
 import org.hibernate.search.mapper.pojo.bridge.binding.impl.BoundTypeBridge;
 import org.hibernate.search.mapper.pojo.bridge.binding.impl.BoundValueBridge;
+import org.hibernate.search.mapper.pojo.bridge.binding.impl.DefaultIdentifierBindingContext;
+import org.hibernate.search.mapper.pojo.bridge.binding.impl.PropertyBindingContextImpl;
+import org.hibernate.search.mapper.pojo.bridge.binding.impl.TypeBindingContextImpl;
+import org.hibernate.search.mapper.pojo.bridge.binding.impl.ValueBindingContextImpl;
 import org.hibernate.search.mapper.pojo.bridge.binding.spi.FieldModelContributor;
+import org.hibernate.search.mapper.pojo.bridge.mapping.impl.BridgeResolver;
 import org.hibernate.search.mapper.pojo.bridge.mapping.programmatic.IdentifierBinder;
 import org.hibernate.search.mapper.pojo.bridge.mapping.programmatic.PropertyBinder;
 import org.hibernate.search.mapper.pojo.bridge.mapping.programmatic.TypeBinder;
 import org.hibernate.search.mapper.pojo.bridge.mapping.programmatic.ValueBinder;
 import org.hibernate.search.mapper.pojo.extractor.ContainerExtractor;
 import org.hibernate.search.mapper.pojo.extractor.impl.BoundContainerExtractorPath;
+import org.hibernate.search.mapper.pojo.extractor.impl.ContainerExtractorBinder;
 import org.hibernate.search.mapper.pojo.extractor.impl.ContainerExtractorHolder;
 import org.hibernate.search.mapper.pojo.extractor.mapping.programmatic.ContainerExtractorPath;
+import org.hibernate.search.mapper.pojo.model.additionalmetadata.building.impl.PojoTypeAdditionalMetadataProvider;
+import org.hibernate.search.mapper.pojo.model.additionalmetadata.impl.PojoEntityTypeAdditionalMetadata;
+import org.hibernate.search.mapper.pojo.model.dependency.impl.PojoPropertyIndexingDependencyConfigurationContextImpl;
+import org.hibernate.search.mapper.pojo.model.dependency.impl.PojoTypeIndexingDependencyConfigurationContextImpl;
+import org.hibernate.search.mapper.pojo.model.impl.PojoModelPropertyRootElement;
+import org.hibernate.search.mapper.pojo.model.impl.PojoModelTypeRootElement;
+import org.hibernate.search.mapper.pojo.model.path.impl.BoundPojoModelPath;
 import org.hibernate.search.mapper.pojo.model.path.impl.BoundPojoModelPathPropertyNode;
 import org.hibernate.search.mapper.pojo.model.path.impl.BoundPojoModelPathTypeNode;
 import org.hibernate.search.mapper.pojo.model.path.impl.BoundPojoModelPathValueNode;
+import org.hibernate.search.mapper.pojo.model.spi.PojoBootstrapIntrospector;
 import org.hibernate.search.mapper.pojo.model.spi.PojoTypeModel;
 
 /**
@@ -45,31 +62,143 @@ import org.hibernate.search.mapper.pojo.model.spi.PojoTypeModel;
  * due to bridges contributing to the index model as we bind them.
  *
  */
-public interface PojoIndexModelBinder {
+public final class PojoIndexModelBinder {
 
-	<T> Optional<BoundPojoModelPathPropertyNode<T, ?>> createEntityIdPropertyPath(PojoTypeModel<T> type);
+	private final BeanResolver beanResolver;
+	private final PojoBootstrapIntrospector introspector;
+	private final ContainerExtractorBinder extractorBinder;
+	private final BridgeResolver bridgeResolver;
+	private final PojoTypeAdditionalMetadataProvider typeAdditionalMetadataProvider;
 
-	<C> BoundContainerExtractorPath<C, ?> bindExtractorPath(
-			PojoTypeModel<C> pojoGenericTypeModel, ContainerExtractorPath extractorPath);
+	PojoIndexModelBinder(MappingBuildContext buildContext,
+			PojoBootstrapIntrospector introspector,
+			ContainerExtractorBinder extractorBinder, BridgeResolver bridgeResolver,
+			PojoTypeAdditionalMetadataProvider typeAdditionalMetadataProvider) {
+		this.beanResolver = buildContext.beanResolver();
+		this.introspector = introspector;
+		this.extractorBinder = extractorBinder;
+		this.bridgeResolver = bridgeResolver;
+		this.typeAdditionalMetadataProvider = typeAdditionalMetadataProvider;
+	}
 
-	<C, V> ContainerExtractorHolder<C, V> createExtractors(
-			BoundContainerExtractorPath<C, V> boundExtractorPath);
+	public <T> Optional<BoundPojoModelPathPropertyNode<T, ?>> createEntityIdPropertyPath(PojoTypeModel<T> type) {
+		Optional<String> entityIdPropertyName = typeAdditionalMetadataProvider
+				.get( type.rawType() )
+				.getEntityTypeMetadata()
+				.flatMap( PojoEntityTypeAdditionalMetadata::getEntityIdPropertyName );
+		if ( !entityIdPropertyName.isPresent() ) {
+			return Optional.empty();
+		}
+		return Optional.of( BoundPojoModelPath.root( type ).property( entityIdPropertyName.get() ) );
+	}
 
-	<I> BoundIdentifierBridge<I> bindIdentifier(
-			Optional<IndexedEntityBindingContext> bindingContext,
-			BoundPojoModelPathPropertyNode<?, I> modelPath, IdentifierBinder binder,
-			Map<String, Object> params);
+	public <C> BoundContainerExtractorPath<C, ?> bindExtractorPath(
+			PojoTypeModel<C> pojoGenericTypeModel, ContainerExtractorPath extractorPath) {
+		return extractorBinder.bindPath( pojoGenericTypeModel, extractorPath );
+	}
 
-	<T> Optional<BoundTypeBridge<T>> bindType(IndexBindingContext bindingContext,
+	public <C, V> ContainerExtractorHolder<C, V> createExtractors(
+			BoundContainerExtractorPath<C, V> boundExtractorPath) {
+		return extractorBinder.create( boundExtractorPath );
+	}
+
+	public <I> BoundIdentifierBridge<I> bindIdentifier(
+			Optional<IndexedEntityBindingContext> indexedEntityBindingContext,
+			BoundPojoModelPathPropertyNode<?, I> modelPath,
+			IdentifierBinder binder, Map<String, Object> params) {
+		PojoTypeModel<I> identifierTypeModel = modelPath.valueWithoutExtractors().getTypeModel();
+
+		IdentifierBinder defaultedBinder = binder;
+		if ( binder == null ) {
+			defaultedBinder = bridgeResolver.resolveIdentifierBinderForType( identifierTypeModel );
+		}
+
+		DefaultIdentifierBindingContext<I> bindingContext = new DefaultIdentifierBindingContext<>(
+				beanResolver,
+				introspector,
+				indexedEntityBindingContext,
+				identifierTypeModel,
+				params
+		);
+
+		return bindingContext.applyBinder( defaultedBinder );
+	}
+
+	public <T> Optional<BoundTypeBridge<T>> bindType(IndexBindingContext indexBindingContext,
 			BoundPojoModelPathTypeNode<T> modelPath, TypeBinder binder,
-			Map<String, Object> params);
+			Map<String, Object> params) {
+		PojoModelTypeRootElement<T> pojoModelRootElement =
+				new PojoModelTypeRootElement<>( modelPath, introspector, typeAdditionalMetadataProvider );
 
-	<P> Optional<BoundPropertyBridge<P>> bindProperty(IndexBindingContext bindingContext,
-			BoundPojoModelPathPropertyNode<?, P> modelPath, PropertyBinder binder, Map<String, Object> params);
+		PojoTypeModel<T> typeModel = modelPath.getTypeModel();
 
-	<V> Optional<BoundValueBridge<V, ?>> bindValue(IndexBindingContext bindingContext,
+		PojoTypeIndexingDependencyConfigurationContextImpl<T> pojoDependencyContext =
+				new PojoTypeIndexingDependencyConfigurationContextImpl<>(
+						introspector,
+						extractorBinder,
+						typeAdditionalMetadataProvider,
+						typeModel
+				);
+
+		TypeBindingContextImpl<T> bindingContext = new TypeBindingContextImpl<>(
+				beanResolver, introspector,
+				typeModel,
+				indexBindingContext,
+				pojoModelRootElement,
+				pojoDependencyContext,
+				params
+		);
+
+		return bindingContext.applyBinder( binder );
+	}
+
+	public <P> Optional<BoundPropertyBridge<P>> bindProperty(IndexBindingContext indexBindingContext,
+			BoundPojoModelPathPropertyNode<?, P> modelPath, PropertyBinder binder, Map<String, Object> params) {
+		PojoModelPropertyRootElement<P> pojoModelRootElement =
+				new PojoModelPropertyRootElement<>( modelPath, introspector, typeAdditionalMetadataProvider );
+		PojoPropertyIndexingDependencyConfigurationContextImpl<P> pojoDependencyContext =
+				new PojoPropertyIndexingDependencyConfigurationContextImpl<>(
+						introspector,
+						extractorBinder,
+						typeAdditionalMetadataProvider,
+						modelPath
+				);
+
+		PojoTypeModel<P> propertyTypeModel = modelPath.getPropertyModel().typeModel();
+
+		PropertyBindingContextImpl<P> bindingContext = new PropertyBindingContextImpl<>(
+				beanResolver, introspector,
+				propertyTypeModel,
+				indexBindingContext,
+				pojoModelRootElement,
+				pojoDependencyContext,
+				params
+		);
+
+		return bindingContext.applyBinder( binder );
+	}
+
+	public <V> Optional<BoundValueBridge<V, ?>> bindValue(IndexBindingContext indexBindingContext,
 			BoundPojoModelPathValueNode<?, ?, V> modelPath, boolean multiValued,
 			ValueBinder binder, Map<String, Object> params,
-			String relativeFieldName, FieldModelContributor contributor);
+			String relativeFieldName, FieldModelContributor contributor) {
+		Integer decimalScale = typeAdditionalMetadataProvider.get( modelPath ).getDecimalScale();
+		IndexFieldTypeDefaultsProvider defaultsProvider = new IndexFieldTypeDefaultsProvider( decimalScale );
 
+		PojoTypeModel<V> valueTypeModel = modelPath.getTypeModel();
+
+		ValueBinder defaultedBinder = binder;
+		if ( binder == null ) {
+			defaultedBinder = bridgeResolver.resolveValueBinderForType( valueTypeModel );
+		}
+
+		ValueBindingContextImpl<V> bindingContext = new ValueBindingContextImpl<>(
+				beanResolver, introspector,
+				valueTypeModel, multiValued,
+				indexBindingContext, defaultsProvider,
+				relativeFieldName, contributor, params
+		);
+
+		return bindingContext.applyBinder( defaultedBinder );
+	}
 }
